+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+## [0.1.3] - 04/01/2020
+
+ - Allow `DataId` to hold `null` IDs and save ID-less models
+ - Fix relevant unit & integration tests
+
 ## [0.1.2] - 03/31/2020
 
-* Fix `also` API in `DataManagerX` extension
-* Improve tests
+ - Fix `also` API in `DataManagerX` extension
+ - Improve tests
 
 ## [0.1.1] - 03/31/2020
 
-* Initial commit
+ - Initial commit